Send tasktracker uninstallation to TaskTracker directly and synchronously again; revert change originally made in opencore r12712.  Since this is a user-initiated action, and its results ought to be immediately apparent to the user in the response, it really must be synchronous.
I don't think it's correct that this deletes the data -- it ought to just uninstall the featurelet on both sides without wiping out the data.  Also, while looking through the original changeset, I can't find any code that actually uses cabochon to uninstall the tasktracker featurelet.  My suspicion is that the "featurelet uninstallation from project" user-initiated action was erroneously confused and merged with the "project deleted, so tasktracker data must also be deleted" case.  In the latter case it *does* actually make sense to delete the *data* from tasktracker asynchronously -- though the tasktracker *featurelet* in zope must still be removed (and removed synchronously) so I bet that's what happened.

But here's the problem: I bet, also, that this removePackage method is called (probably behind the scenes, but maybe explicitly) during project deletion as well as featurelet uninstallation.  If this is the case, then tasktracker will be getting this "delete data" signal twice on project deletion -- one synchronously and one asynchronously via cabochon.  This shouldn't cause any visible errors, since opencore never (to my knowledge at least) listens for a signal back from tasktracker via cabochon, so even if that second deletion request to tasktracker fails spectacularly badly, that failure will never make it back to the user.  (Hopefully, though, tasktracker will just gracefully ignore the request.)  The worst that might happen is that cabochon won't get back a "good" signal and will repeat the delete request ad infinitum ... which will both consume unnecessary cycles (both for cabochon and for tasktracker) ... though *this* could, theoretically, one day get back to the user ... if the project is reinstated and new task data is created, and cabochon, which is still blindly trying to delete the damn thing, tells tasktracker one final time to wipe the project's data.  That will finally satisfy cabochon, but the user will suddenly be data-less.

So, there are two next steps:
 * Investigate whether tasktracker will explode on a second "delete" request or will gracefully ignore it. Make sure it gracefully ignores it.
 * Really this whole thing points to important missing functionality for featurelets: different behavior on "uninstallation" and "deletion." Make this possible.



git-svn-id: https://svn.socialplanning.org/svn/opencore/plugins/oc-tt/trunk@16203 40f73b84-6177-4a8e-8c61-7eee4b1974ed